fram_totalizer_tv_based

+++ Core/Inc/interrupt_codes.h
@@ -91,7 +91,7 @@ extern uint32_t  r_volTotaliser1,
 				 r_volTotaliser2,
 		         old_r_volTotaliser2;
 
-extern double running_volTotaliser1_array[4],
+extern float running_volTotaliser1_array[4],
 			 running_volTotaliser1c_array[4],
 			 running_volTotaliser2_array[4],
 			 running_volTotaliser2c_array[4];
